requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
redis
psycopg2-binary
drf-yasg
gunicorn
whitenoise
